backend/requirements.txt
# 核心依赖
flask==2.3.3
flask-cors==4.0.1
python-dotenv==1.0.0
requests==2.31.0
python-pptx==1.0.2
openai==1.38.0
httpx==0.26.0

# 生产环境
gunicorn==21.2.0

# 开发工具
pytest==8.3.2
flake8==7.1.0